+++ tests/rolling_restart/rolling_restart_suite_test.go
@@ -22,6 +22,7 @@ var (
 	namespace  = "test-rolling-restart"
 	dcName     = "dc2"
 	dcYaml     = "../testdata/default-single-rack-2-node-dc.yaml"
+	taskYaml   = "../testdata/tasks/rolling_restart.yaml"
 	dcResource = fmt.Sprintf("CassandraDatacenter/%s", dcName)
 	ns         = ginkgo_util.NewWrapper(testName, namespace)
 )
@@ -48,7 +49,7 @@ func TestLifecycle(t *testing.T) {
 
 var _ = Describe(testName, func() {
 	Context("when in a new cluster", func() {
-		Specify("the operator can perform a rolling restart", func() {
+		Specify("operator is installed and cluster is created", func() {
 			By("deploy cass-operator with kustomize")
 			err := kustomize.Deploy(namespace)
 			Expect(err).ToNot(HaveOccurred())
@@ -61,9 +62,11 @@ var _ = Describe(testName, func() {
 
 			ns.WaitForDatacenterReady(dcName)
 
-			step = "trigger restart"
+		})
+		Specify("the operator can perform a rolling restart with rollingRestartRequested spec change", func() {
+			step := "trigger restart"
 			json := `{"spec": {"rollingRestartRequested": true}}`
-			k = kubectl.PatchMerge(dcResource, json)
+			k := kubectl.PatchMerge(dcResource, json)
 			ns.ExecAndLog(step, k)
 
 			// Ensure we actually set the condition
@@ -84,6 +87,22 @@ var _ = Describe(testName, func() {
 			Expect(ns.OutputAndLog(step, k)).To(Equal("true true"))
 
 			ns.WaitForDatacenterReady(dcName)
+		})
+		Specify("cassandratask can be used to do rolling restart of the cluster", func() {
+			step := "creating a cassandra task to do rolling restart"
+			k := kubectl.ApplyFiles(taskYaml)
+			ns.ExecAndLog(step, k)
+
+			// Wait for the task to be completed
+			ns.WaitForCompleteTask("rolling-restart")
+
+			// Verify each pod does have the annotation..
+			json := `jsonpath={.items[0].metadata.annotations.control\.k8ssandra\.io/restartedAt}`
+			k = kubectl.Get("pods").
+				WithLabel(fmt.Sprintf("cassandra.datastax.com/datacenter=%s", dcName)).
+				WithFlag("field-selector", "status.phase=Running").
+				FormatOutput(json)
+			ns.WaitForOutputPatternAndLog(step, k, `^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$`, 360)
 		})
 	})
 })